requirements (1).txt
streamlit
opencv-python-headless
Pillow
scikit-image
